pyproject.toml
[build-system]
requires = ["setuptools>=63.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ihec-scrubber"
version = "0.1.0"
description = "IHEC Mega Scrubber for polling station data"
requires-python = ">=3.10"
authors = [
    { name = "Hamlet Unified Ops" }
]
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.2",
    "lxml>=4.9.3",
    "pytesseract>=0.3.10",
    "pdf2image>=1.17.0",
    "Pillow>=9.5.0",
    "selenium>=4.10.0",
    "webdriver-manager>=4.0.0",
    "python-dotenv>=1.0.0",
    "boto3>=1.28.0",
]

[project.optional-dependencies]
test = ["pytest>=7.4.0"]

[tool.pytest.ini_options]
addopts = "-ra"
pythonpath = ["."]
